Add some basic reporting to the filter
This allows the filter to report on how many messages were included
or excluded, which is useful for verifying the number of messages
against AWS CloudWatch. If a JSON decoding error is encountered,
the report also allows the user to determine how far through the
input file the tool got before the error was encountered.