Add helpers for replaying events into Matomo
The provided Dockerfile uses a clone of a fork of the Matomo
analytics script, capable of handling `X-Forwarded-For` headers.
This, and a pre-defined access log are copied into the image.
A template command to replay events is echoed into the terminal -
a Matomo password currently has to be entered in plaintext, so should
not be baked into the Docker image.
The container is then spun up for the developer to enter the command.

This setup is clearly not ideal. But it is functional.
On the basis that something is better than nothing, it should be
iterated on as improvements become clearer.