Improve error status responses when downloading
If a user has decided not to continue, then terminating the script
after having completing vacuously should not result in a non-zero
exit code. That is, completing should be successful.
Further, by setting the `-e` flag, there is no need to explicitly
exit with the code from the AWS CLI call - a non-zero exit code will
automatically cause the script to terminate and yield that code.