Copy script from `verify-matomo-tool`
A new script was added to a separate repo. The script contains an
improved method for retrieving logs from AWS CloudWatch via Insights.
This commit includes the contents of the repo to consolidate scripts
into a single location. (Only one commit has been made in the other
repo, so no history has been lost by copying in this manner.)

Note that some minor changes have been included as part of the copy.