Delete existing download and filter directories
The existing method for downloading logs via the AWS CLI is being
removed. In its place, the retrieval process added in the parent
commit should be used instead. This downloads logs via AWS CloudWatch
Insights, so the query syntax is much improved.

The filter process was only needed for massaging the results from the
CLI download. With the new process no such massaging is required.