Extract some constants; simplify the date input
The date format is used in a few places. Note that the filename
format has here been changed slightly to make the formats consistent.
The filename suffix is only used once, but never changes. Both are
now extracted as constants and elevated to the top of the script for
visibility. The name of a variable tracking the output filename has
also been lowercased so that it is not assumed to be a constant.

Additionally, the input variable called START_DATE had previously
required a timestamp. For simplicity, this is now reduced to a date
as the name suggests, and the timestamp is assumed to be the start of
the day.