Switch to the logging library
Currently, messages are being printed. This is fine if all output is
always wanted. However, during automated runs, it is often helpful to
be verbose to assist with debugging any potential problems. So being
able to print at different levels can be beneficial. Instead of
printing based on whether or not a debug flag is provided, the
logging library allows all logs to be created the same way, but only
printed if an appropriate flag is specified. Thus developers only
need to see verbose output when they want, but pipelines can specify
an appropriate parameter for suitably verbose output every run.

Currently, the log level can be overridden by environment variables,
in a similar fashion to the existing script parameters.