Add more logging
This should allow an automated platform to output more verbose
information about the steps it is taking, such as the periods it
requests, how many requests are written for each period, and notably
whether the script is hanging while waiting for a response to
complete. This last is especially important, because if nothing is
printed, it can be hard to know what has caused the script to stall.