Simplify loop logic
Instead of looping for a day, for each of the days specified, it
should be equivalent to simply determine the entire length of the
period, and continue looping until all requests in that period have
been downloaded.

Simplifying the loop in this way allows us to eliminate both an
unnecessary level of nesting, and an additional final iteration that
required checking outside the loop and thus duplicated the code
sending the requests.

The size of the period to loop over has been extracted into a
constant to increase its visibility. This is a temporary change until
an environment variable is established for it.