Make period width overridable
This should mean adjustments to the period width can happen either in
code (if the default value should be changed) or through environment
variables (if a single run of the script is safe to use non-default
values for).

Since it is now possible to provide large values, the warning about
receiving too many requests is now an error and aborts the script.
This is also better behaviour for automation, since the process
running the script should also fail, and the error should propagate.