Add script to retrieve logs from a specified day
The included script should make automating this process simpler. It
passes on the appropriate environment variables to the python script.
Note that some optional variables have purposely not been named, but
Docker should pass any values set even without the explicit
assignment. Explicit assignments have been retained for some to use
the built-in error-if-not-set bash functionality. If bash cannot be
used for automation, this may need further refinement.

Note also that both local helper scripts now also ensure that the
shell is in the correct directory for the Docker build process, to
avoid obscure bugs where the Docker context is not as expected.
This, again, currently relies on bash so may not extend to a fully
automated process.