Switch from bash to sh
Docker in docker does not have bash, so the helper scripts need to be
Bourne Shell friendly. The only thing that wasn't already compatible
was a reference to the current directory, which has been swapped out
for a hopefully-equivalent compatible version.

Note that the new version is taken from a SO answer. The code is
therefore available under the MIT licence already. However it does
require attribution, and guidance from [1] has been followed.

[1]: https://stackoverflow.blog/2009/06/25/attribution-required/